requirements2.txt
imutils==0.5.4
numpy==1.23.4
opencv_python==4.7.0.72
pandas==1.5.1
Pillow==9.4.0
torch==1.13.1
